+++ packages/nextjs/src/app-router/client/ClerkProvider.tsx
@@ -57,7 +57,7 @@ export const ClientClerkProvider = (props: NextClerkProviderProps) => {
       return new Promise(res => {
         window.__clerk_internal_invalidateCachePromise = res;
         startTransition(() => {
-          //@ts-expect-error next exitsts on window
+          //@ts-expect-error next exists on window
           if (window.next?.version && typeof window.next.version === 'string' && window.next.version.startsWith('13')) {
             router.refresh();
           } else {