+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @clerk/ui
 
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies [[`1273b04ec`](https://github.com/clerk/javascript/commit/1273b04ecf1866b59ef59a74abe31dbcc726da2c), [`b8acf30a7`](https://github.com/clerk/javascript/commit/b8acf30a7c2e3a484f20f3586e8316fc1d3cc390), [`c44925501`](https://github.com/clerk/javascript/commit/c449255011ad65aa0a409facb90fce46b5f2b6be)]:
+  - @clerk/types@4.6.1
+  - @clerk/elements@0.8.0
+  - @clerk/clerk-react@5.2.5
+  - @clerk/shared@2.3.1
+
 ## 0.1.1
 
 ### Patch Changes